extract long description trimming logic to work for comments too, but putting it in CommonTextFromParts instead
